mil: store elements in Cell's
For interior mutability.

In other words, this allows transformation algos (`crate::xform`) to
hold a slice of the program and perform mutations on it at the same
time, at the cost of ever allowing multithreading.

This cost shouldn't be too problematic, as it's going to be easy to keep
each function non-shared while still having per-function parallelism.