mil: change insns to have fixed-size results
The set of opcodes in MIL has been changed so that each opcodes has a
specific, constant size for its (integer) output register, regardless
of inputs.

This is intended to simplify many algorithms operating on the SSA
repr, especially around type checking, introducing widening/narrowing
instructions (not yet implemented, but I see it on the horizon).